requirements.txt
fastapi
uvicorn[standard]

# OpenTelemetry (required for your server.py)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi

# Your app dependencies
python-dotenv
pydantic
httpx
requests

# GitHub API (optional but likely used by your portfolio scanner)
PyGithub

numpy
scikit-learn
google-generativeai>=0.4.0
